feat(acquisitions): rename purchasing to acquisitions and implement new components
- Updated Navbar to reflect the change from "Purchasing" to "Acquisitions".
- Added LotsTable component for displaying lots with pagination and search functionality.
- Introduced ModalAddLot component for adding new lots with form validation.
- Created ModalAddSupplier component for adding new suppliers with form validation.
- Implemented SuppliersTable component for displaying suppliers with pagination and search functionality.
- Developed main acquisitions page to manage lots and suppliers, including modals for adding new entries.
- Added utility functions for form validation and initial values for lots and suppliers.
- Updated Dashboard to link to the new acquisitions page.
