+++ src/app/acquisitions/LotsTable.jsx
