+++ src/app/acquisitions/ModalAddLot.jsx
@@ -11,7 +11,7 @@ import { Formik, Form, Field, useFormikContext } from 'formik'
 import getClassPrefixer from '~/app/UI/classPrefixer'
 import TextField from '~/app/UI/Shared/FormikTextField'
 import SelectField from '~/app/UI/Shared/FormikSelect'
-import { getAddLotValidationSchema, getAddLotInitialValues } from '~/app/purchasing/utils'
+import { getAddLotValidationSchema, getAddLotInitialValues } from '~/app/acquisitions/utils'
 import apiFetch from '~/app/Lib/apiFetch'
 import { useToken } from '~/app/store/useToken'
 import { useState } from 'react'
