+++ src/app/acquisitions/utils.js
